Add Next.js prerender (ISR) to mainpage
